requirements_dev.txt
flask
cvu-python
onnx
onnxruntime
Pillow
pytest
black
flake8